dash testing method. folder structure changed
