+++ vizro-core/hatch.toml
@@ -64,8 +64,8 @@ schema-check = ["python schemas/generate.py --check"]
 test = "pytest tests --headless {args}"
 test-e2e-component-library = "pytest tests/e2e/component_library/test_component_library.py --headless {args}"
 test-e2e-vizro-dom-elements = [
-  "gunicorn dashboard:app -b 0.0.0.0:5002 -w 3 --timeout 90 &",
-  "tests/e2e/vizro/dashboards/wait-for-it.sh 127.0.0.1:5002 -t 30",
+  "gunicorn dashboard:app -b 0.0.0.0:5002 -w 1 --timeout 90 &",
+  "tests/tests_utils/e2e/vizro/dashboards/wait-for-it.sh 127.0.0.1:5002 -t 30",
   "pytest tests/e2e/vizro/test_dom_elements --headless {args}"
 ]
 test-integration = "pytest tests/integration --headless {args}"
