+++ vizro-core/tests/e2e/vizro/custom_components/custom_charts/bar_custom.py
