+++ vizro-core/tests/e2e/vizro/custom_components/custom_components/new_dropdown.py
