+++ vizro-core/tests/e2e/vizro/custom_components/custom_components/range_slider_non_cross.py
