+++ vizro-core/tests/e2e/vizro/test_dom_elements/test_charts.py
@@ -1,11 +1,7 @@
-import e2e.constants as cnst
-from e2e.checkers import check_text
-from e2e.helpers import webdriver_click_waiter, webdriver_waiter
-from e2e.paths import href_path, nav_card_text_path
+import e2e.vizro.constants as cnst
 
 
-def test_modebar(dash_br_driver):
+def test_modebar(dash_br):
     """Check that modebar element exist for the chart."""
-    check_text(dash_br_driver, xpath=href_path(href=cnst.HOME_PAGE_PATH), text=cnst.HOME_PAGE)
-    webdriver_click_waiter(dash_br_driver, xpath=nav_card_text_path(href=cnst.FILTERS_PAGE_PATH))
-    webdriver_waiter(dash_br_driver, xpath="//*[@class='modebar-container']")
+    dash_br.multiple_click(f"a[href='{cnst.FILTERS_PAGE_PATH}']", 1)
+    dash_br.wait_for_element(".modebar-container")
