+++ vizro-core/tests/e2e/vizro/test_dom_elements/test_filters.py
@@ -1,84 +1,58 @@
-import e2e.constants as cnst
+import e2e.vizro.constants as cnst
 import pytest
-from e2e.checkers import (
-    check_graph_is_loading,
-    check_range_slider_value,
-    check_slider_value,
-)
-from e2e.helpers import graph_load_waiter, webdriver_click_waiter, webdriver_waiter
-from e2e.navigation import page_select
-from e2e.paths import (
-    checklist_value_path,
-    dropdown_arrow_path,
-    dropdown_clear_all_path,
-    dropdown_value_path,
-    kpi_card_text_path,
-    radio_items_value_path,
-    range_slider_value_path,
-    slider_value_path,
-)
+from e2e.vizro.checkers import check_graph_is_loading, check_slider_value
+from e2e.vizro.navigation import page_select, select_dropdown_value
+from e2e.vizro.paths import categorical_components_value_path, slider_value_path
+from e2e.vizro.waiters import graph_load_waiter
 from hamcrest import assert_that, equal_to
-from selenium.webdriver.common.by import By
 
 
-def test_dropdown(dash_br_driver):
-    page_select(dash_br_driver, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_clear_all_path())
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_arrow_path())
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_value_path(value=1))
-    check_graph_is_loading(dash_br_driver, graph_id=cnst.SCATTER_GRAPH_ID)
+def test_dropdown(dash_br):
+    page_select(dash_br, page_path=cnst.FILTERS_PAGE_PATH, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
+    select_dropdown_value(dash_br, value=2)
+    check_graph_is_loading(dash_br, graph_id=cnst.SCATTER_GRAPH_ID)
 
 
-def test_radio_items(dash_br_driver):
-    page_select(dash_br_driver, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
-    webdriver_click_waiter(dash_br_driver, xpath=radio_items_value_path(title="Species", value=2))
-    check_graph_is_loading(dash_br_driver, graph_id=cnst.SCATTER_GRAPH_ID)
+@pytest.mark.parametrize(
+    "filter_id",
+    [cnst.CHECK_LIST_FILTER_FILTERS_PAGE, cnst.RADIO_ITEMS_FILTER_FILTERS_PAGE],
+    ids=["checklist", "radio_items"],
+)
+def test_categorical_filters(dash_br, filter_id):
+    page_select(dash_br, page_path=cnst.FILTERS_PAGE_PATH, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
+    dash_br.multiple_click(categorical_components_value_path(elem_id=filter_id, value=2), 1)
+    check_graph_is_loading(dash_br, graph_id=cnst.SCATTER_GRAPH_ID)
 
 
-def test_checklist(dash_br_driver):
-    page_select(dash_br_driver, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
-    webdriver_click_waiter(dash_br_driver, xpath=checklist_value_path(title="Species", value=2))
-    check_graph_is_loading(dash_br_driver, graph_id=cnst.SCATTER_GRAPH_ID)
-
-
-def test_slider(dash_br_driver):
-    page_select(dash_br_driver, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
-    webdriver_click_waiter(
-        dash_br_driver,
-        slider_value_path(elem_id=cnst.SLIDER_FILTER_FILTERS_PAGE, value=2),
-    )
-    check_graph_is_loading(dash_br_driver, graph_id=cnst.SCATTER_GRAPH_ID)
-    check_slider_value(dash_br_driver, num="0.6", elem_id=cnst.SLIDER_FILTER_FILTERS_PAGE)
+def test_slider(dash_br):
+    page_select(dash_br, page_path=cnst.FILTERS_PAGE_PATH, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
+    dash_br.multiple_click(slider_value_path(elem_id=cnst.SLIDER_FILTER_FILTERS_PAGE, value=2), 1)
+    check_graph_is_loading(dash_br, graph_id=cnst.SCATTER_GRAPH_ID)
+    check_slider_value(dash_br, expected_end_value="0.6", elem_id=cnst.SLIDER_FILTER_FILTERS_PAGE)
 
 
 @pytest.mark.xfail(reason="Should be fixed later in vizro by Petar")
 # Right now is failing with the next error:
 # AssertionError: Element number is '4', but expected number is '4.3'
-def test_range_slider(dash_br_driver):
-    page_select(dash_br_driver, page_name=cnst.FILTERS_PAGE)
-    graph_load_waiter(dash_br_driver, graph_id=cnst.SCATTER_GRAPH_ID)
-    webdriver_click_waiter(
-        dash_br_driver,
-        range_slider_value_path(elem_id=cnst.RANGE_SLIDER_FILTER_FILTERS_PAGE, value=4),
-    )
-    check_graph_is_loading(dash_br_driver, graph_id=cnst.SCATTER_GRAPH_ID)
-    check_range_slider_value(
-        dash_br_driver, left_num="4.3", right_num="7", elem_id=cnst.RANGE_SLIDER_FILTER_FILTERS_PAGE
+def test_range_slider(dash_br):
+    page_select(dash_br, page_path=cnst.FILTERS_PAGE_PATH, page_name=cnst.FILTERS_PAGE, graph_id=cnst.SCATTER_GRAPH_ID)
+    dash_br.multiple_click(slider_value_path(elem_id=cnst.RANGE_SLIDER_FILTER_FILTERS_PAGE, value=4), 1)
+    check_graph_is_loading(dash_br, graph_id=cnst.SCATTER_GRAPH_ID)
+    check_slider_value(
+        dash_br, elem_id=cnst.RANGE_SLIDER_FILTER_FILTERS_PAGE, expected_start_value="4.3", expected_end_value="7"
     )
 
 
-def test_dropdown_homepage(dash_br_driver):
-    graph_load_waiter(dash_br_driver, graph_id=cnst.AREA_GRAPH_ID)
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_clear_all_path())
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_arrow_path())
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_value_path(value=1))
-    check_graph_is_loading(dash_br_driver, cnst.AREA_GRAPH_ID)
+def test_dropdown_homepage(dash_br):
+    graph_load_waiter(dash_br, graph_id=cnst.AREA_GRAPH_ID)
+    select_dropdown_value(dash_br, value=2)
+    check_graph_is_loading(dash_br, cnst.AREA_GRAPH_ID)
 
 
-def test_dropdown_kpi_indicators_page(dash_br_driver):
-    page_select(dash_br_driver, page_name=cnst.KPI_INDICATORS_PAGE)
-    webdriver_waiter(dash_br_driver, xpath=kpi_card_text_path())
-    values = dash_br_driver.find_elements(by=By.XPATH, value=kpi_card_text_path())
+def test_dropdown_kpi_indicators_page(dash_br):
+    page_select(dash_br, page_path=cnst.KPI_INDICATORS_PAGE_PATH, page_name=cnst.KPI_INDICATORS_PAGE)
+    dash_br.wait_for_text_to_equal(".card-body", "73902")
+    values = dash_br.find_elements(".card-body")
     values_text = [value.text for value in values]
     assert_that(
         actual_or_assertion=values_text,
@@ -96,11 +70,9 @@ def test_dropdown_kpi_indicators_page(dash_br_driver):
             ]
         ),
     )
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_clear_all_path())
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_arrow_path())
-    webdriver_click_waiter(dash_br_driver, xpath=dropdown_value_path(value=2))
-    webdriver_waiter(dash_br_driver, xpath=kpi_card_text_path())
-    values = dash_br_driver.find_elements(by=By.XPATH, value=kpi_card_text_path())
+    select_dropdown_value(dash_br, value=2)
+    dash_br.wait_for_text_to_equal(".card-body", "67434")
+    values = dash_br.find_elements(".card-body")
     values_text = [value.text for value in values]
     assert_that(
         actual_or_assertion=values_text,
