--- vizro-core/tests/tests_utils/e2e/navigation.py
@@ -1,20 +1,0 @@
-import time
-
-from e2e.helpers import graph_load_waiter, webdriver_click_waiter, webdriver_waiter
-from e2e.paths import accordion_path, page_accordion_path, page_title_path
-
-
-def accordion_select(chromdriver, accordion_name):
-    webdriver_click_waiter(chromdriver, accordion_path(accordion_name))
-    webdriver_waiter(chromdriver, f'//*{accordion_path(accordion_name)}[@aria-expanded="true"]')
-
-
-def page_select(driver, page_name, graph_id=None, accordion_name=None):
-    if accordion_name:
-        # choose accordion
-        accordion_select(driver, accordion_name=accordion_name.upper())
-    time.sleep(1)  # to let accordion open
-    webdriver_click_waiter(driver, page_accordion_path(page_name))
-    webdriver_waiter(driver, page_title_path(page_name))
-    if graph_id:
-        graph_load_waiter(driver, graph_id)
