+++ vizro-core/tests/tests_utils/e2e/vizro/constants.py
@@ -23,12 +23,15 @@ FILTERS_TAB_CONTAINER = "filters tab_container"
 FILTERS_COMPONENTS_CONTAINER = "filters components container"
 SCATTER_GRAPH_ID = "Scatter"
 BOX_GRAPH_ID = "box graph"
+RADIO_ITEMS_FILTER_FILTERS_PAGE = "radio-filters-page"
+CHECK_LIST_FILTER_FILTERS_PAGE = "check filters-page"
 SLIDER_FILTER_FILTERS_PAGE = "slider-filters-page"
 RANGE_SLIDER_FILTER_FILTERS_PAGE = "range slider-filters-page"
 
 PARAMETERS_PAGE = "parameters_p@ge! (tabs + containers)"
 PARAMETERS_PAGE_PATH = "/parameters_page"
 PARAMETERS_TAB_CONTAINER = "parameters-tab_container"
+PARAMETERS_SUB_TAB_ID = "sub tab params"
 PARAMETERS_SUB_TAB_CONTAINER_ONE = "parameters sub t@b container 1"
 PARAMETERS_SUB_TAB_CONTAINER_TWO = "parameters sub t@b container 2"
 PARAMETERS_CONTAINER = "parameters container"
@@ -37,6 +40,7 @@ BAR_GRAPH_ID_CONTAINER = "b@R-graph__container"
 HISTOGRAM_GRAPH_ID = "-histogram-graph--"
 
 KPI_INDICATORS_PAGE = "kpi indicators page"
+KPI_INDICATORS_PAGE_PATH = "/kpi-indicators-page"
 
 EXPORT_PAGE = "export page"
 EXPORT_PAGE_PATH = "/exportp"
@@ -49,6 +53,7 @@ TABLE_POP_RANGE_ID = "table pop range"
 TABLE_POP_DATE_ID = "table pop date"
 
 TABLE_AG_GRID_PAGE = "table AG grid page"
+TABLE_AG_GRID_PAGE_PATH = "/table-ag-grid-page"
 TABLE_AG_GRID_ID = "123_ag_grid_table"
 BOX_AG_GRID_PAGE_ID = "B@x on ag grid page"
 TABLE_AG_GRID_CONTAINER = "table_ag_grid_container"
