+++ vizro-core/tests/tests_utils/e2e/vizro/waiters.py
@@ -1,0 +1,3 @@
+def graph_load_waiter(driver, graph_id):
+    """Waiting for graph's x-axis to appear."""
+    driver.wait_for_element(f"div[id='{graph_id}'] path[class='xtick ticks crisp']")
